fetch/metadata/generated/form-submission-sec-fetch-site-redirect.https.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Site - redirection</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(method, key, path) {
    const win = window.open();
    const inputs = [
        ['key', key],
        ['body', `<script>opener.postMessage('done', '*');</${''}script>`],
        ...new URL(path, location.origin).searchParams
      ]
      .map(([name, value]) =>
        `<input type="hidden" name="${name}" value="${value}">`
      )
      .join('\n');

    return new Promise((resolve) => {
        addEventListener('message', function(event) {
          if (event.source === win) {
            resolve();
          }
        });
        win.document.open();
        win.document.write(`
          <!DOCTYPE html>
          <form method="${method}" action="${path}">${inputs}</form>
          <script>
            const form = document.getElementsByTagName('form')[0];
            form.submit();
          </${''}script>
        `);
        win.document.close();
      })
      .then(() => win.close());
  }

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py';
    const key = '{{uuid()}}';
    return induceRequest('GET', key, 'https://{{hosts[alt][]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' +path)
      .then(() => fetch(path + '?retrieve&key=' + key))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Cross-Site -> Same-Origin');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py';
    const key = '{{uuid()}}';
    return induceRequest('GET', key, 'https://{{hosts[alt][]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{domains[www2]}}:{{ports[https][0]}}' +path)
      .then(() => fetch(path + '?retrieve&key=' + key))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Cross-Site -> Same-Site');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py';
    const key = '{{uuid()}}';
    return induceRequest('GET', key, 'https://{{hosts[alt][]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{hosts[alt][]}}:{{ports[https][0]}}' +path)
      .then(() => fetch(path + '?retrieve&key=' + key))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Cross-Site -> Cross-Site');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py';
    const key = '{{uuid()}}';
    return induceRequest('GET', key, 'https://{{host}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{hosts[alt][]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' +path)
      .then(() => fetch(path + '?retrieve&key=' + key))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin');

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py';
    const key = '{{uuid()}}';
    return induceRequest('GET', key, 'https://{{host}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=http://{{host}}:{{ports[http][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' +path)
      .then(() => fetch(path + '?retrieve&key=' + key))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade');
  </script>
  </body>
</html>
